refactor: CardEditorScreen 컴포저블 구조 개선 (#154)
- ui/CardEditorBottomScreen.kt을 상위 패키지로 이동하고 CardEditorScreen으로 명칭 변경
- AssetBrowserPanelContent를 CardEditorContent로 통합하여 불필요한 계층 제거
- selectedTab 상태를 CardEditorScreen으로 상향 조정 (State Hoisting)
- ControlHeader, SpawnedObjectRow 등 내부 컴포저블 가시성을 `private`으로 제한
